feat(user): admin can reset password for given user (#292)
## Context <!-- ie. explanations, background, documentation -->

This PR introduces ability for admin to reset password of user. This is
currently implemented only for users that have no email. The FE makes
sure that in case a temporary password already exists for a user, it can
be shown to the admin, otherwise the Reset Password button will be
shown. This is implemented in
https://github.com/aula-app/aula-frontend/pull/874

In the future, we should implement
https://github.com/aula-app/aula-frontend/issues/800 for users with
registered email. This PR offers flexibility for that future
implementation.

## Checklist

- [x] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [x] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [x] Changelist updated
- ~[] Backward and forward compatible with FE~ ** this BE PR must be
released before this FE PR is released:
https://github.com/aula-app/aula-frontend/pull/874**
- [x] Doesn't need update in the database <!-- If it does, please
describe how to deploy it without downtime -->
- [ ] Must be deployed ASAP (HOTFIX)