fix: Message->addMessage not working when using user hash id on target (#303)
New messages were not being create when frontend used user ids on the
message target. The logic for checking if a hash id or id was being
passed was wrong.

It is already being done inside the checkUserId, checkGroupId and
checkRoomId methods.

Fixes
https://github.com/aula-app/aula-frontend/issues/890#issue-3470216362

## Context <!-- ie. explanations, background, documentation -->

<!-- Example:
After the last update it became apparent that we did fetch the new
results, but didn't actually store them properly. This caused a glitch
in the UI whenever the user would refresh the page.
-->

## Checklist

- [x] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [x] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [x] Changelist updated
- [x] Backward and forward compatible with FE <!-- If not, please
describe in detail and include other PR links -->
- [x] Doesn't need update in the database <!-- If it does, please
describe how to deploy it without downtime -->
- [x] Must be deployed ASAP (HOTFIX)

---------

Co-authored-by: nikola-maric-aula <nikola.maric@aula.de>