Fix race condition when multiple roles update (#360)
## Context <!-- ie. explanations, background, documentation -->

Fix #359 

when multiple request to addUserRole or deleteUserRole where being made
to the same user, a race condition was happening and the sequential
requests were getting the old roles value before the update for earlier
requests. The solution was using a transaction and also the FOR UPDATE
in the select query to get the roles, without that just a transaction
didn't solve the issue.

<!-- Example:
After the last update it became apparent that we did fetch the new
results, but didn't actually store them properly. This caused a glitch
in the UI whenever the user would refresh the page.
-->

## Checklist

- [x] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [x] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [ ] Changelist updated
- [x] Backward and forward compatible with
[aula-frontend/releases](https://github.com/aula-app/aula-frontend/releases)
<!-- If not, please describe in detail and include other PR links -->
- [x] Doesn't need update in the database <!-- If it does, please
describe how to deploy it without downtime -->
- [x] Must be deployed ASAP (HOTFIX)
- [ ] Needs update of [docs.aula.de](https://docs.aula.de/)
([repo](https://github.com/leonard-haas/docs_aula)) <!-- If it does,
please ping Leonard OR include link to the change in the docs repo -->

---------

Co-authored-by: nikola-maric-aula <nikola.maric@aula.de>